Add `make help`

+++ Makefile
@@ -8,7 +8,7 @@ VENV_DEV_PATH := .venvs/dev
 VENV_RELEASE_PATH := .venvs/release
 
 help:
-	@echo "Usage: make TARGET [ANOTHER_TARGET]..."
+	@grep -E '^[a-zA-Z_-]+:.*?## .*$$' $(MAKEFILE_LIST) | sort | awk 'BEGIN {FS = ":.*?## "}; {printf "\033[36m%-30s\033[0m %s\n", $$1, $$2}'
 
 _venv_dev:
 	virtualenv --version >/dev/null || pip install --user virtualenv
@@ -23,40 +23,40 @@ _venv_release:
 	. "${VENV_RELEASE_PATH}/bin/activate"
 	pip install --upgrade pip setuptools wheel twine
 
-flake8: _venv_dev
+flake8: _venv_dev ## Run flake8 for code analysis
 	flake8 kong/
 
-mypy: _venv_dev
+mypy: _venv_dev ## Run mypy for code static type checking
 	mypy kong/
 
-install:
+install: ## (Re)install from source tree, user-wide
 	pip install --user --force-reinstall .
 	### smoke check after user-wide install ###
 	kong-incubator
 
-build: _venv_release
+build: _venv_release ## Build Python dists ready for `publish_`
 	### sanity check before building ###
 	pip install .
 	python setup.py clean --all bdist_wheel sdist
 
-publish_testpypi: _venv_release
+publish_testpypi: _venv_release ## Publish the `build` dists to test.pypi.org
 	twine upload --repository-url https://test.pypi.org/legacy/ dist/*
 
-publish_pypi: _venv_release
+publish_pypi: _venv_release ## Publish the `build` dists to pypi.org
 	twine upload dist/*
 
-test: _venv_dev
+test: _venv_dev ## Start docker-compose and run tests against it
 	git submodule update --init --recursive
 	docker-compose --file testkong/docker-compose.yml up --detach \
 		kong-database kong-migration kong
 	KADMIN_SECRET="k0n64dm1n" pytest --cov --spec --instafail --diff-type=auto
 
-testdown:
+testdown: ## Remove docker-compose services and volumes
 	docker-compose --file testkong/docker-compose.yml down --volumes
 
-all: test build
+next: test build ## Run `test`s, if passing `build` dists
 
-clean: testdown
+clean: testdown ## Purge testenv, .venvs, dists, and tool caches
 	rm -rf dist build *.egg-info kong/__pycache__
 	rm -rf .venvs
 	rm -rf .pytest_cache .mypy_cache
